Refactored remaining files in the src folder
